% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemoPack-package.R
\docType{package}
\name{DemoPack-package}
\alias{DemoPack}
\alias{DemoPack-package}
\title{DemoPack: A Demo Package to Show How Easy Creating an R Package is!}
\description{
This package was created as part of a demonstration for an R-Club presentation 2026-01-30.
}
\author{
\strong{Maintainer}: Aubrey DuBois \email{aubreydubois1@icloud.com} (\href{https://orcid.org/0000-0003-3666-5883}{ORCID})

}
\keyword{internal}
