% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_ids.R
\name{standardize_ids}
\alias{standardize_ids}
\title{Standardizes IDs to match main data repository format}
\usage{
standardize_ids(df, var)
}
\arguments{
\item{df}{dataframe containing ID variable}

\item{var}{name of the ID variable}
}
\value{
A dataframe with a new variable column "std_id" that contains the standardized IDs
}
\description{
Standardizes IDs to match main data repository format
}
\examples{
sample_data <- as.data.frame(list(
  id = c("123"," 0 123","00-123","0123","000456789","``3456","3456"),
  name = c(
  "ABC Incorporated","AlphaBravoCharlie Inc.","abc inc","AlfaBravoCharlie Incorp",
  "DEF Company","ghi services","ghi sercives")
  ), stringsAsFactors = FALSE)

standardize_ids(sample_data, id)

}
