\name{demopack-package}
\alias{demopack-package}
\alias{demopack}
\docType{package}
\title{
Mean Standard Deviation}
\description{
Returns mean and standard deviation of data set of 100 random numbers
}
\details{
\tabular{ll}{
Package: \tab demopack\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-22\cr
License: \tab What license is it under?\cr
}
}
\author{
Yen <yen@jhu.edu>
}
\references{
}

\keyword{ package }
\seealso{
}
\examples{
}
